:art: simplied cogs and added load_cogs method
