+++ examples/basic_cogs/run.py
@@ -5,13 +5,8 @@ from cogs import OnReadyCog, SayCog, ErrorHandler
 
 class Bot(Client):
     def __init__(self, *args, **kwargs):
-        self.load_cogs()
+        super().load_cogs(OnReadyCog, SayCog, ErrorHandler)
         super().__init__(*args, **kwargs)
-
-    def load_cogs(self):
-        self.load_cog(OnReadyCog)
-        self.load_cog(SayCog)
-        self.load_cog(ErrorHandler)
 
 
 if __name__ == "__main__":
